src/main/java/ru/netology/domian/Radio.java
package ru.netology.domian;

public class Radio {
    private String name;
    private int minVolume;
    private int currentVolume;
    private int maxVolume = 10;
    private int minRadioStation;
    private int currentRadioStation;
    private int maxRadioStation = 9;
    private boolean on;

    public String getName() {

        return name;
    }

    public void setName(String name) {

        this.name = name;
    }

    public int getMinVolume() {

        return minVolume;
    }

    public void setMinVolume(int minVolume) {

        this.minVolume = minVolume;
    }

    public int getMaxVolume() {

        return maxVolume;
    }

    public void setMaxVolume(int maxVolume) {

        this.maxVolume = maxVolume;
    }

    public int getCurrentVolume() {

        return currentVolume;
    }

    public void setCurrentVolume(int currentVolume) {

        this.currentVolume = currentVolume;
    }

    public int getMinRadioStation() {

        return minRadioStation;
    }

    public void setMinRadioStation(int minRadioStation) {

        this.minRadioStation = minRadioStation;
    }

    public int getMaxRadioStation() {

        return maxRadioStation;
    }

    public void setMaxRadioStation(int maxRadioStation) {

        this.maxRadioStation = maxRadioStation;
    }

    public int getCurrentRadioStation() {

        return currentRadioStation;
    }

    public boolean isOn() {

        return on;
    }

    public void setOn(boolean on) {

        this.on = on;
    }

    public void increaseVolume() {

        if (currentVolume < maxVolume) {
            currentVolume++;
        } else {
            currentVolume = minVolume;
        }
    }

    public void decreaseVolume() {

        if (currentVolume > minVolume) {
            currentVolume--;
        } else {
            currentVolume = maxVolume;
        }
    }

    public void prevRadioStation() {

        if (currentRadioStation > minRadioStation) {
            currentRadioStation--;
        } else {
            currentRadioStation = maxRadioStation;
        }
    }

    public void nextRadioStation() {

        if (currentRadioStation < maxRadioStation) {
            currentRadioStation++;
        } else {
            currentRadioStation = minRadioStation;
        }
    }

    public void setCurrentRadioStation(int currentRadioStation) {

        if (currentRadioStation > maxRadioStation) {
            this.currentRadioStation = maxRadioStation;
            return;
        }
        if (currentRadioStation < minRadioStation) {
            this.currentRadioStation = minRadioStation;
            return;
        }
        this.currentRadioStation = currentRadioStation;
    }
}
